pyproject.toml
[tool.poetry]
name = "beachPatrol"
version = "0.1.0"
description = "A discord bot that patrols the beach."
authors = ["Drew&Pate <drew@lido.fi>"]
readme = "README.md"
packages = [
    { include = "beachPatrol", from = "src" }
]


[tool.poetry.dependencies]
python = "^3.11"
twspace-dl = "^2023.7.24.1"
openai = "^0.28.1"
langchain = "^0.0.306"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
py-cord = "^2.4.1"
celery = "^5.3.4"
redis = "^5.0.1"
beautifulsoup4 = "^4.12.3"
pytube = "^15.0.0"
yt-dlp = "^2024.8.6"
newspaper4k = "^0.9.3.1"
lxml = {extras = ["html-clean"], version = "^5.3.0"}
nltk = "3.9"
pymupdf = "^1.24.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
